Moved long raiden.ts code blocks to helper.ts functions
