+++ raiden-ts/tests/e2e/raiden.spec.ts
@@ -111,7 +111,7 @@ describe('Raiden', () => {
   });
 
   test('create from other params and RaidenState', async () => {
-    expect.assertions(8);
+    expect.assertions(7);
 
     // token address not found as an account in provider
     await expect(Raiden.create(provider, token, storage, contractsInfo, config)).rejects.toThrow(
